rendersource+plane_master+alpha flags

+++ OpenDreamShared/Dream/DreamFilter.cs
@@ -49,7 +49,7 @@ public sealed record DreamFilterAlpha : DreamFilter {
     [ViewVariables, DataField("y")] public float Y;
     [ViewVariables, DataField("icon")] public int Icon; // Icon resource ID
     [ViewVariables, DataField("render_source")] public string RenderSource; // String that gets special processing in the render loop
-    [ViewVariables, DataField("flags")] public float Flags;
+    [ViewVariables, DataField("flags")] public short Flags;
 }
 
 [Serializable, NetSerializable]
